+++ R/A_swGenericMethods.R
@@ -137,7 +137,7 @@ check_version <- function(
       expected[[c(1, 3)]] <- 0
     }
 
-    if (level %in% "major") {
+    if (identical(level, "major")) {
       # zero the minor-level
       has[[c(1, 2)]] <- 0
       expected[[c(1, 2)]] <- 0
@@ -230,7 +230,7 @@ format_timestamp <- function(object) {
 #'         `"SWA"` added as `outkey` 8 for a new total of 30
 #'
 #' @examples
-#'   x <- sw_upgrade(rSOILWAT2::sw_exampleData, verbose = TRUE)
+#' x <- sw_upgrade(rSOILWAT2::sw_exampleData, verbose = TRUE)
 #'
 #' @md
 #' @exportMethod sw_upgrade